+++ lib/src/catalog/catalog-return-objects.dart
@@ -125,14 +125,16 @@ class BatchRetrieveCatalogObjectsResponse extends ApiResponse {
               .map((error) => SquareError.fromJson(error))
               .toList()
           : null,
-      objects: json['objects'] != null ?
-          (json['objects'] as List).map((item) => CatalogObject.fromJson(item)).toList()
-          :null ,
-      relatedObjects: json['related_objects'] != null 
-        ?(json['related_objects'] as List)
-            .map((item) => CatalogObject.fromJson(item))
-            .toList()
-        : null,
+      objects: json['objects'] != null
+          ? (json['objects'] as List)
+              .map((item) => CatalogObject.fromJson(item))
+              .toList()
+          : null,
+      relatedObjects: json['related_objects'] != null
+          ? (json['related_objects'] as List)
+              .map((item) => CatalogObject.fromJson(item))
+              .toList()
+          : null,
     );
   }
 }