+++ lib/src/employees/employees-objects.dart
@@ -65,7 +65,7 @@ class Employee {
 
   Map<String, dynamic> toJson() {
     var body = Map<String, dynamic>();
-    
+
     if (id != null) body['id'] = id;
     if (firstName != null) body['first_name'] = firstName;
     if (lastName != null) body['last_name'] = lastName;
@@ -75,7 +75,7 @@ class Employee {
     if (status != null) body['status'] = getStringFromEmployeeStatus(status);
     if (createdAt != null) body['created_at'] = createdAt.toString();
     if (updatedAt != null) body['updated_at'] = updatedAt.toString();
-    
+
     return body;
   }
 }